backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.18
python-dotenv==1.0.1
httpx==0.28.1
rapidfuzz==3.10.1
playwright==1.49.0
greenlet==3.1.1
apscheduler==3.10.4

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
